Поправить замечания, сделать рефакторинг тестов
