+++ src/services/TaskManager.java
@@ -5,7 +5,6 @@ import tasks.Subtask;
 import tasks.Task;
 
 import java.util.List;
-import java.util.TreeSet;
 
 public interface TaskManager {
     List<Task> getTasksList();
@@ -48,5 +47,5 @@ public interface TaskManager {
 
     List<Task> getHistory();
 
-    TreeSet<Task> getPrioritizedTaskList();
+    List<Task> getPrioritizedTasks();
 }
